requirements.txt
astor==0.8.1
astroid==2.11.7
asttokens==2.0.5
autograd==1.4
autograd-gamma==0.5.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.31.2
formulaic==0.3.4
future==0.18.2
idna==3.3
imbalanced-learn==0.9.0
interface-meta==1.3.0
ipykernel==6.12.1
ipython==8.2.0
isort==5.10.1
jedi==0.18.1
joblib==1.1.0
JPype1==1.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
lifelines==0.27.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
nest-asyncio==1.5.5
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pylint==2.14.5
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
stack-data==0.2.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.1
torch==1.11.0
torchvision==0.12.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
wrapt==1.14.1
